Changes after rebase plus refactor
