+++ packages/framework-integration-tests/src/constants.ts
@@ -5,5 +5,3 @@ export const beforeHookException = 'Before hook throwing exception'
 export const beforeHookMutationID = 'mutation-but-with-input-changes'
 export const beforeHookMutationIDModified = beforeHookMutationID + '-modified'
 export const beforeHookQuantity = 21
-export const throwExceptionId = 'throw-exception-id'
-export const beforeHookException = 'Before hook throwing exception'
