+++ packages/framework-types/src/concepts/filter-hooks.ts
@@ -1,8 +1,20 @@
 import { UserEnvelope } from '../envelope'
 import { CommandInput } from './command'
+import { FilterFor } from '../searcher'
+import { Class } from '../typelevel'
+import { ReadModelInterface } from './read-model'
 
 export interface CommandFilterHooks {
   readonly before?: Array<CommandBeforeFunction>
 }
 
 export type CommandBeforeFunction = (input: CommandInput, currentUser?: UserEnvelope) => CommandInput
+
+export interface ReadModelFilterHooks {
+  readonly before?: Array<ReadModelBeforeFunction>
+}
+
+export type ReadModelBeforeFunction = (
+  filter: FilterFor<Class<ReadModelInterface>>,
+  currentUser?: UserEnvelope
+) => FilterFor<Class<ReadModelInterface>>
