+++ packages/framework-types/src/concepts/read-model.ts
@@ -1,8 +1,6 @@
-import { RoleAccess, UUID } from '.'
+import { ReadModelFilterHooks, RoleAccess, UUID } from '.'
 import { Class } from '../typelevel'
 import { PropertyMetadata } from 'metadata-booster'
-import { FilterFor } from '../searcher'
-import { UserEnvelope } from '../envelope'
 
 export interface ReadModelInterface {
   id: UUID
@@ -18,12 +16,3 @@ export interface ReadModelMetadata {
   readonly authorizedRoles: RoleAccess['authorize']
   readonly before: NonNullable<ReadModelFilterHooks['before']>
 }
-
-export interface ReadModelFilterHooks {
-  readonly before?: Array<BeforeFunction>
-}
-
-export type BeforeFunction = (
-  filter: FilterFor<Class<ReadModelInterface>>,
-  currentUser?: UserEnvelope
-) => FilterFor<Class<ReadModelInterface>>
